fist

+++ src/main/java/by/ntck/sten/model/SkladH.java
@@ -31,7 +31,7 @@ public class SkladH {
 	private String operthiya;
 
 	@Column(name = "kol_vo")
-	private float kol_vo;
+	private double kol_vo;
 
 	@Column(name = "data_oper")
 	private String data_oper;
@@ -67,7 +67,7 @@ public class SkladH {
 	private String is_del;
 
 	@Column(name = "kol_vo_old")
-	private int kol_vo_old;
+	private double kol_vo_old;
 
 	@Column(name = "obrabotano")
 	private String obrabotano;
@@ -136,11 +136,11 @@ public class SkladH {
 		this.operthiya = operthiya;
 	}
 
-	public float getKol_vo() {
+	public double getKol_vo() {
 		return kol_vo;
 	}
 
-	public void setKol_vo(float kol_vo) {
+	public void setKol_vo(double kol_vo) {
 		this.kol_vo = kol_vo;
 	}
 
@@ -232,11 +232,11 @@ public class SkladH {
 		this.is_del = is_del;
 	}
 
-	public int getKol_vo_old() {
+	public double getKol_vo_old() {
 		return kol_vo_old;
 	}
 
-	public void setKol_vo_old(int kol_vo_old) {
+	public void setKol_vo_old(double kol_vo_old) {
 		this.kol_vo_old = kol_vo_old;
 	}
